Update PlayerOneCamera SDK to v3.2.0 and more (see changelog)
