+++ debian/indi-playerone/changelog
@@ -1,3 +1,12 @@
+indi-playerone (1.2) bionic; urgency=low
+
+  * Add single camera driver for multi threading
+  * Add flip property with bayer compensation
+  * Add nickname property for multi camera driver
+  * Update PlayerOneCamera SDK to v3.2.0
+
+ -- Hiroshi SAITO <hiro3110g@gmail.com>  Fri, 24 Feb 2023 12:00:00 +0900
+
 indi-playerone (1.1) bionic; urgency=low
 
   * Update libplayerone with PlayerOnePW SDK v1.1.0
