+++ debian/libplayerone/changelog
@@ -1,3 +1,9 @@
+libplayerone (1.0) bionic; urgency=low
+
+  * Update PlayerOneCamera SDK to v3.2.0
+
+ -- Hiroshi SAITO <hiro3110g@gmail.com>  Thu, 24 Fri 2023 08:00:00 +0900
+
 libplayerone (0.9) bionic; urgency=low
 
   * Update PlayerOnePW SDK to v1.1.0
@@ -6,7 +12,7 @@ libplayerone (0.9) bionic; urgency=low
 
 libplayerone (0.8) bionic; urgency=low
 
-  * Add PlayerOne Filter Wheel driver
+  * Add PlayerOnePW SDK v1.0.0
   * Update PlayerOneCamera SDK to v3.1.1
 
  -- Hiroshi SAITO <hiro3110g@gmail.com>  Wed, 18 Jan 2023 08:00:00 +0900
