+++ indi-playerone/CMakeLists.txt
@@ -13,7 +13,7 @@ find_package(USB1 REQUIRED)
 find_package(Threads REQUIRED)
 
 set(PLAYERONE_VERSION_MAJOR 1)
-set(PLAYERONE_VERSION_MINOR 1)
+set(PLAYERONE_VERSION_MINOR 2)
 
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config.h.cmake ${CMAKE_CURRENT_BINARY_DIR}/config.h )
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/indi_playerone.xml.cmake ${CMAKE_CURRENT_BINARY_DIR}/indi_playerone.xml)
@@ -36,6 +36,7 @@ endif()
 
 ########### indi_playerone_ccd ###########
 set(indi_playerone_SRCS
+   ${CMAKE_CURRENT_SOURCE_DIR}/playerone_base.cpp
    ${CMAKE_CURRENT_SOURCE_DIR}/playerone_ccd.cpp
    )
 
@@ -43,6 +44,18 @@ add_executable(indi_playerone_ccd ${indi_playerone_SRCS})
 target_link_libraries(indi_playerone_ccd ${INDI_LIBRARIES} ${CFITSIO_LIBRARIES} ${PLAYERONE_LIBRARIES} ${USB1_LIBRARIES} ${ZLIB_LIBRARY} ${CMAKE_THREAD_LIBS_INIT})
 if (HAVE_WEBSOCKET)
     target_link_libraries(indi_playerone_ccd ${Boost_LIBRARIES})
+endif()
+
+########### indi_playerone_single_ccd ###########
+set(indi_playerone_single_SRCS
+   ${CMAKE_CURRENT_SOURCE_DIR}/playerone_base.cpp
+   ${CMAKE_CURRENT_SOURCE_DIR}/playerone_single_ccd.cpp
+   )
+
+add_executable(indi_playerone_single_ccd ${indi_playerone_single_SRCS})
+target_link_libraries(indi_playerone_single_ccd ${INDI_LIBRARIES} ${CFITSIO_LIBRARIES} ${PLAYERONE_LIBRARIES} ${USB1_LIBRARIES} ${ZLIB_LIBRARY} ${CMAKE_THREAD_LIBS_INIT})
+if (HAVE_WEBSOCKET)
+    target_link_libraries(indi_playerone_single_ccd ${Boost_LIBRARIES})
 endif()
 
 ########### indi_playerone_wheel ###########
@@ -67,11 +80,13 @@ ENDIF()
 
 if (CMAKE_SYSTEM_PROCESSOR MATCHES "arm*")
 target_link_libraries(indi_playerone_ccd rt)
+target_link_libraries(indi_playerone_single_ccd rt)
 target_link_libraries(indi_playerone_wheel rt)
 target_link_libraries(playerone_camera_test rt)
 endif (CMAKE_SYSTEM_PROCESSOR MATCHES "arm*")
 
 install(TARGETS indi_playerone_ccd RUNTIME DESTINATION bin)
+install(TARGETS indi_playerone_single_ccd RUNTIME DESTINATION bin)
 install(TARGETS indi_playerone_wheel RUNTIME DESTINATION bin)
 install(TARGETS playerone_camera_test RUNTIME DESTINATION bin)
 install(FILES ${CMAKE_CURRENT_BINARY_DIR}/indi_playerone.xml DESTINATION ${INDI_DATA_DIR})
