+++ libplayerone/99-player_one_astronomy.rules
