+++ libplayerone/libplayerone.spec
@@ -87,10 +87,12 @@ make DESTDIR=%{buildroot} install
 %license libplayerone/license.txt
 
 %changelog
+* Fri Feb 24 2023 Hiroshi Saito <hiro3110g@gmail.com>
+- update PlayerOneCamera SDK to v3.2.0
 * Sun Feb 05 2023 Hiroshi Saito <hiro3110g@gmail.com>
 - update PlayerOnePW SDK to v1.1.0
 * Wed Jan 18 2023 Hiroshi Saito <hiro3110g@gmail.com>
-- add PlayerOne Filter Wheel driver
+- add PlayerOnePW SDK v1.0.0
 - update PlayerOneCamera SDK to v3.1.1
 * Thu Dec 15 2022 Hiroshi Saito <hiro3110g@gmail.com>
 - update PlayerOneCamera SDK to v3.1.0
