+++ libplayerone/license.txt
